Implémentation de toutes les fonctionnalités d'affichage, création et modification des livres.

+++ controllers/bookController.js
@@ -1,25 +1,80 @@
 const Book = require('../models/Book');
+const fs = require('fs');
 
 exports.createBook = (req, res, next) => {
-    delete req.body._id;
+    const bookObject = JSON.parse(req.body.book);
+    delete bookObject._id;
+    delete bookObject._userId;
     const book = new Book({
-        ...req.body,
+        ...bookObject,
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${
+            req.file.filename
+        }`,
     });
     book.save()
-        .then(() => res.status(201).json({ message: 'Livre enregistré !' }))
-        .catch((error) => res.status(400).json({ error: error }));
+        .then(() => {
+            res.status(201).json({ message: 'Objet enregistré !' });
+        })
+        .catch((error) => {
+            res.status(400).json({ error: error });
+        });
 };
 
 exports.modifyBook = (req, res, next) => {
-    Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet modifié !' }))
-        .catch((error) => res.status(400).json({ error: error }));
+    const bookObject = req.file
+        ? {
+              ...JSON.parse(req.body.book),
+              imageUrl: `${req.protocol}://${req.get('host')}/images/${
+                  req.file.filename
+              }`,
+          }
+        : { ...req.body };
+
+    delete bookObject._userId;
+    Book.findOne({ _id: req.params.id })
+        .then((book) => {
+            if (book.userId != req.auth.userId) {
+                res.status(400).json({ message: 'Not authorized' });
+            } else {
+                Book.updateOne(
+                    { _id: req.params.id },
+                    { ...bookObject, _id: req.params.id }
+                )
+                    .then(() =>
+                        res.status(200).json({ message: 'Objet modifié !' })
+                    )
+                    .catch((error) => {
+                        res.status(401).json({ error: error });
+                    });
+            }
+        })
+        .catch((error) => {
+            res.status(401).json({ error: error });
+        });
 };
 
 exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
-        .catch((error) => res.status(400).json({ error: error }));
+    Book.findOne({ _id: req.params.id })
+        .then((book) => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Non autorisé' });
+            } else {
+                const filename = book.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Book.deleteOne({ _id: req.params.id })
+                        .then(() =>
+                            res
+                                .status(200)
+                                .json({ message: 'Objet supprimé !' })
+                        )
+                        .catch((error) =>
+                            res.status(401).json({ error: error })
+                        );
+                });
+            }
+        })
+        .catch((error) => res.status(500).json({ error: error }));
 };
 
 exports.getOneBook = (req, res, next) => {
